Move SyncDataAccessor to 'util' module
