+++ ethereum/core/src/main/java/tech/pegasys/teku/core/signatures/SlashingProtector.java
@@ -23,6 +23,7 @@ import tech.pegasys.teku.bls.BLSPublicKey;
 import tech.pegasys.teku.data.signingrecord.ValidatorSigningRecord;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.util.file.SyncDataAccessor;
 
 public class SlashingProtector {
